Source code release
Appifyer is still considered to be in beta and has not been updated in
over a year. However, it is fully functioning, more stable and better
looking than previous alpha (and earlier) versions. I am now releasing
the source code for all fellow coders to see and use. If you have
modifications or suggestions, feel free to comment or contribute.